+++ lib/creq/project.rb
@@ -1,0 +1,25 @@
+# encoding: UTF-8
+
+require_relative 'settings'
+
+module Creq
+  module Project
+    extend self
+
+    def init(directory = Dir.pwd)
+      Dir.chdir(directory) {
+        folders.each {|f| Dir.mkdir(f) unless Dir.exist?(f)}
+      }
+      Settings.save
+    end
+
+    def folders
+      [Settings.bin,
+       Settings.assets,
+       Settings.src,
+       Settings.lib,
+       Settings.kbd,
+       Settings.tt]
+    end
+  end
+end