+++ test/cli/cli_toc_spec.rb
@@ -21,7 +21,7 @@ describe "creq doc" do
   it 'must print toc' do
     skip() # TODO how to check output of Thor Cli.start?
     inside_sandbox do
-      inside_req { File.write('req.md', content) }
+      inside_src { File.write('req.md', content) }
       Cli.start(["toc", "--skipid='cmp1'"])
     end
   end